moving pages to a separate folder
