+++ src/main.js
@@ -8,9 +8,9 @@ import * as components from 'vuetify/components'
 import * as directives from 'vuetify/directives'
 
 import { createRouter, createWebHistory } from 'vue-router'
-import MainPage from './components/MainPage.vue'
-import MoviePage from './components/MoviePage.vue'
-import BookmarksPage from './components/BookmarksPage.vue'
+import MainPage from './pages/MainPage.vue'
+import MoviePage from './pages/MoviePage.vue'
+import BookmarksPage from './pages/BookmarksPage.vue'
 import { createPinia } from 'pinia'
 
 const vuetify = createVuetify({
